+++ src/components/Inputs/Button.tsx
@@ -6,7 +6,8 @@ import { Shadow } from 'react-native-shadow-2';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 import { colord, extend } from 'colord';
 import namesPlugin from 'colord/plugins/names';
-import { propsMerger } from '../../internalUtils/propsMerger';
+import type { MayBeFunction } from '../../internalUtils/defaultProps';
+import { propsMerger, useGetDefaultProps } from '../../internalUtils/defaultProps';
 import type { OmitKey } from '../../internalUtils/types';
 import { useTheme } from '../../main/theme';
 import { mLoading } from '../Modals/mLoading';
@@ -76,16 +77,20 @@ export interface ButtonProps<FunRtn extends void | any | Promise<any> = unknown>
   shadowProps?: ShadowProps;
 }
 
-export interface ButtonPropsTheme extends Partial<OmitKey<ButtonProps, 'testID' | 'nativeID' | 'text' | 'children'>> {}
+/** If a function, it's run as a React Hook. */
+export type ButtonPropsTheme = MayBeFunction<Partial<OmitKey<ButtonProps, 'testID' | 'nativeID' | 'text' | 'children'>>>;
+
 
 export function Button<T extends(void | any | Promise<any>)>(props: ButtonProps<T>): JSX.Element {
-  const { colors, fonts, props: themeProps } = useTheme();
+  const { colors, fonts, defaultProps: themeProps } = useTheme();
+
+  const defaultProps = useGetDefaultProps(themeProps.Button);
 
   // ts strange error on stretch = strechRow if not doing this here separatedly
   const mergedProps = useMemo(() => propsMerger<ButtonProps>({
-    props: [themeProps.Button, props],
+    props: [defaultProps, props],
     stylesKeys: ['style', 'textStyle', 'iconContainerStyle', 'containerStyle'],
-  }), [props, themeProps.Button]);
+  }), [defaultProps, props]);
 
   const {
     marginTop: marginTopArg = false,