+++ src/internalUtils/defaultProps.tsx
@@ -1,8 +1,9 @@
 import { StyleSheet } from 'react-native';
+import type { Obj } from './types';
 
 
 
-
+/** Merges the props and the indicated styles. */
 export function propsMerger<T extends Record<string, any>>({ props, stylesKeys = [] }: {
   props: (Partial<T> | null | undefined)[];
   stylesKeys?: (keyof T)[];
@@ -16,7 +17,13 @@ export function propsMerger<T extends Record<string, any>>({ props, stylesKeys =
   };
 }
 
+/** If defaultProp it's a function, then returns its return. */
+export function useGetDefaultProps<T extends Obj>(param: T | (() => T)): T {
+  return typeof param === 'function' ? param() : param;
+}
 
+/** For T, returns T | (() => T) */
+export type MayBeFunction<T> = T | (() => T);
 
 // export function usePropsMerger<T extends Record<string, unknown>>({ props, stylesKeys }: {
 //   props: (T | null | undefined)[];